Add parameter for vectorises to addDyad
